Rename servlet common root package
