+++ instrumentation/servlet/servlet-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/servlet/v3_0/AsyncDispatchAdvice.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.servlet.v3_0;
 
-import static io.opentelemetry.javaagent.instrumentation.servlet.ServletHelper.CONTEXT_ATTRIBUTE;
+import static io.opentelemetry.javaagent.instrumentation.servlet.common.ServletHelper.CONTEXT_ATTRIBUTE;
 
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.context.Context;
