+++ instrumentation/servlet/servlet-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/servlet/common/AgentServletInstrumenterBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.servlet;
+package io.opentelemetry.javaagent.instrumentation.servlet.common;
 
 import static java.util.Collections.emptyList;
